@@ -3,7 +3,7 @@ name = "elspeth"
 version = "0.1.0"
 description = "Auditable Sense/Decide/Act pipelines for high-reliability systems"
 readme = "README.md"
-requires-python = ">=3.11"
+requires-python = ">=3.12"
 license = { text = "MIT" }
 authors = [{ name = "John" }]
 keywords = ["elspeth", "audit", "pipeline", "sda", "traceability", "data-processing"]
@@ -12,8 +12,8 @@ classifiers = [
     "Intended Audience :: Developers",
     "License :: OSI Approved :: MIT License",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
 
@@ -26,6 +26,7 @@ dependencies = [
     "dynaconf>=3.2",
     "pydantic>=2.6",
     "python-dotenv>=1.0",  # Load .env files for secrets
+    "pyyaml>=6.0",  # YAML parsing for configuration files
 
     # Plugin system
     "pluggy>=1.4",
@@ -78,6 +79,7 @@ dev = [
     # Type stubs
     "pandas-stubs>=2.2",
     "types-networkx>=3.2",
+    "types-PyYAML>=6.0",
     # Mutation testing (Phase 3)
     "mutmut>=2.4",
 ]
@@ -119,7 +121,7 @@ packages = ["src/elspeth"]
 
 # === Ruff Configuration ===
 [tool.ruff]
-target-version = "py311"
+target-version = "py312"
 line-length = 140
 src = ["src", "tests"]
 
@@ -156,7 +158,7 @@ known-first-party = ["elspeth"]
 
 # === Mypy Configuration ===
 [tool.mypy]
-python_version = "3.11"
+python_version = "3.12"
 strict = true
 warn_return_any = true
 warn_unused_configs = true
