@@ -1,5 +1,5 @@
-# src/elspeth/plugins/llm/aimd_throttle.py
-"""AIMD (Additive Increase, Multiplicative Decrease) throttle for LLM API calls.
+# src/elspeth/plugins/pooling/throttle.py
+"""AIMD (Additive Increase, Multiplicative Decrease) throttle for API calls.
 
 Implements TCP-style congestion control:
 - On capacity error: multiply delay (fast ramp down)
