@@ -1,30 +1,49 @@
 """Tests for AzurePromptShield transform."""
 
 from typing import TYPE_CHECKING, Any
-from unittest.mock import Mock
+from unittest.mock import MagicMock, Mock
 
 import pytest
 
 from elspeth.plugins.config_base import PluginConfigError
 
 if TYPE_CHECKING:
-    pass
+    from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
 
 
-def make_mock_context(http_response: dict[str, Any] | None = None) -> Mock:
-    """Create mock PluginContext with HTTP client."""
+def make_mock_context() -> Mock:
+    """Create mock PluginContext for testing."""
     from elspeth.plugins.context import PluginContext
 
-    ctx = Mock(spec=PluginContext, run_id="test-run")
+    return Mock(spec=PluginContext, run_id="test-run")
 
-    if http_response is not None:
-        response_mock = Mock()
-        response_mock.status_code = 200
-        response_mock.json.return_value = http_response
-        response_mock.raise_for_status = Mock()
-        ctx.http_client.post.return_value = response_mock
 
-    return ctx
+def make_prompt_shield_with_mock_response(
+    config: dict[str, Any],
+    response_data: dict[str, Any],
+) -> tuple["AzurePromptShield", MagicMock]:
+    """Create Prompt Shield transform with mocked HTTP client.
+
+    Returns the transform and the mock client for assertions.
+    """
+    from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+    transform = AzurePromptShield(config)
+
+    # Create mock response
+    response_mock = MagicMock()
+    response_mock.status_code = 200
+    response_mock.json.return_value = response_data
+    response_mock.raise_for_status = MagicMock()
+
+    # Create mock client
+    mock_client = MagicMock()
+    mock_client.post.return_value = response_mock
+
+    # Inject mock client directly (bypassing _get_http_client)
+    transform._http_client = mock_client
+
+    return transform, mock_client
 
 
 class TestAzurePromptShieldConfig:
@@ -173,24 +192,20 @@ class TestAzurePromptShieldTransform:
 
     def test_clean_content_passes(self) -> None:
         """Content without attacks passes through."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        transform, _ = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": False},
                 "documentsAnalysis": [{"attackDetected": False}],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         row = {"prompt": "What is the weather?", "id": 1}
         result = transform.process(row, ctx)
 
@@ -199,24 +214,20 @@ class TestAzurePromptShieldTransform:
 
     def test_user_prompt_attack_returns_error(self) -> None:
         """User prompt attack detection returns error."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        transform, _ = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": True},
                 "documentsAnalysis": [{"attackDetected": False}],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         row = {"prompt": "Ignore previous instructions", "id": 1}
         result = transform.process(row, ctx)
 
@@ -228,24 +239,20 @@ class TestAzurePromptShieldTransform:
 
     def test_document_attack_returns_error(self) -> None:
         """Document attack detection returns error."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        transform, _ = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": False},
                 "documentsAnalysis": [{"attackDetected": True}],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         row = {"prompt": "Summarize this document", "id": 1}
         result = transform.process(row, ctx)
 
@@ -255,24 +262,20 @@ class TestAzurePromptShieldTransform:
 
     def test_both_attacks_detected(self) -> None:
         """Both attack types can be detected simultaneously."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        transform, _ = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": True},
                 "documentsAnalysis": [{"attackDetected": True}],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         row = {"prompt": "Malicious content", "id": 1}
         result = transform.process(row, ctx)
 
@@ -296,13 +299,16 @@ class TestAzurePromptShieldTransform:
             }
         )
 
-        ctx = make_mock_context()
-        ctx.http_client.post.side_effect = httpx.HTTPStatusError(
+        # Create mock client that raises HTTPStatusError
+        mock_client = MagicMock()
+        mock_client.post.side_effect = httpx.HTTPStatusError(
             "Rate limited",
             request=Mock(),
             response=Mock(status_code=429),
         )
+        transform._http_client = mock_client
 
+        ctx = make_mock_context()
         row = {"prompt": "test", "id": 1}
         result = transform.process(row, ctx)
 
@@ -327,9 +333,12 @@ class TestAzurePromptShieldTransform:
             }
         )
 
-        ctx = make_mock_context()
-        ctx.http_client.post.side_effect = httpx.RequestError("Connection failed")
+        # Create mock client that raises RequestError
+        mock_client = MagicMock()
+        mock_client.post.side_effect = httpx.RequestError("Connection failed")
+        transform._http_client = mock_client
 
+        ctx = make_mock_context()
         row = {"prompt": "test", "id": 1}
         result = transform.process(row, ctx)
 
@@ -341,24 +350,20 @@ class TestAzurePromptShieldTransform:
 
     def test_skips_missing_configured_field(self) -> None:
         """Transform skips fields not present in the row."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        transform, _ = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt", "optional_field"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": False},
                 "documentsAnalysis": [{"attackDetected": False}],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         # Row is missing "optional_field"
         row = {"prompt": "safe prompt", "id": 1}
         result = transform.process(row, ctx)
@@ -367,31 +372,27 @@ class TestAzurePromptShieldTransform:
 
     def test_skips_non_string_fields(self) -> None:
         """Transform skips non-string field values."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        transform, mock_client = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt", "count"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": False},
                 "documentsAnalysis": [{"attackDetected": False}],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         # count is an int, should be skipped
         row = {"prompt": "safe prompt", "count": 42, "id": 1}
         result = transform.process(row, ctx)
 
         assert result.status == "success"
         # Only one API call should be made (for "prompt" field)
-        assert ctx.http_client.post.call_count == 1
+        assert mock_client.post.call_count == 1
 
     def test_malformed_api_response_returns_error(self) -> None:
         """Malformed API responses return error (fail-closed security posture).
@@ -400,20 +401,18 @@ class TestAzurePromptShieldTransform:
         garbage, we must not let potentially malicious content pass through
         undetected. Malformed responses are treated as errors, not "no attack".
         """
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        # Mock a malformed response (missing expected fields)
+        transform, _ = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt"],
                 "schema": {"fields": "dynamic"},
-            }
+            },
+            {"unexpectedField": "value"},
         )
 
-        # Mock a malformed response (missing expected fields)
-        ctx = make_mock_context({"unexpectedField": "value"})
-
+        ctx = make_mock_context()
         row = {"prompt": "test", "id": 1}
         result = transform.process(row, ctx)
 
@@ -430,25 +429,21 @@ class TestAzurePromptShieldTransform:
         If documentsAnalysis is missing from the response, that's a malformed
         response that should be rejected, not treated as "no document attack".
         """
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        # Mock a response with only userPromptAnalysis (documentsAnalysis missing)
+        transform, _ = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        # Mock a response with only userPromptAnalysis (documentsAnalysis missing)
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": False},
                 # documentsAnalysis missing
-            }
+            },
         )
 
+        ctx = make_mock_context()
         row = {"prompt": "test", "id": 1}
         result = transform.process(row, ctx)
 
@@ -474,13 +469,16 @@ class TestAzurePromptShieldTransform:
             }
         )
 
-        ctx = make_mock_context()
-        ctx.http_client.post.side_effect = httpx.HTTPStatusError(
+        # Create mock client that raises HTTPStatusError with 400
+        mock_client = MagicMock()
+        mock_client.post.side_effect = httpx.HTTPStatusError(
             "Bad Request",
             request=Mock(),
             response=Mock(status_code=400),
         )
+        transform._http_client = mock_client
 
+        ctx = make_mock_context()
         row = {"prompt": "test", "id": 1}
         result = transform.process(row, ctx)
 
@@ -492,47 +490,38 @@ class TestAzurePromptShieldTransform:
 
     def test_all_fields_mode_scans_all_string_fields(self) -> None:
         """When fields='all', all string fields are scanned."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        transform, mock_client = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": "all",
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": False},
                 "documentsAnalysis": [{"attackDetected": False}],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         # Row with multiple string fields plus non-string
         row = {"prompt": "safe", "title": "also safe", "count": 42, "id": 1}
         result = transform.process(row, ctx)
 
         assert result.status == "success"
         # Should have called API twice (for "prompt" and "title", not "count" or "id")
-        assert ctx.http_client.post.call_count == 2
+        assert mock_client.post.call_count == 2
 
     def test_multiple_documents_analysis(self) -> None:
         """Document attack is detected if any document shows attack."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        # Second document shows attack
+        transform, _ = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com",
                 "api_key": "test-key",
                 "fields": ["prompt"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        # Second document shows attack
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": False},
                 "documentsAnalysis": [
@@ -540,9 +529,10 @@ class TestAzurePromptShieldTransform:
                     {"attackDetected": True},
                     {"attackDetected": False},
                 ],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         row = {"prompt": "test", "id": 1}
         result = transform.process(row, ctx)
 
@@ -552,30 +542,26 @@ class TestAzurePromptShieldTransform:
 
     def test_api_called_with_correct_endpoint_and_headers(self) -> None:
         """API is called with correct endpoint URL and headers."""
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
+        transform, mock_client = make_prompt_shield_with_mock_response(
             {
                 "endpoint": "https://test.cognitiveservices.azure.com/",
                 "api_key": "my-secret-key",
                 "fields": ["prompt"],
                 "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = make_mock_context(
+            },
             {
                 "userPromptAnalysis": {"attackDetected": False},
                 "documentsAnalysis": [{"attackDetected": False}],
-            }
+            },
         )
 
+        ctx = make_mock_context()
         row = {"prompt": "test prompt", "id": 1}
         transform.process(row, ctx)
 
         # Verify the API call
-        ctx.http_client.post.assert_called_once()
-        call_args = ctx.http_client.post.call_args
+        mock_client.post.assert_called_once()
+        call_args = mock_client.post.call_args
 
         # Check URL (trailing slash should be stripped)
         expected_url = "https://test.cognitiveservices.azure.com/contentsafety/text:shieldPrompt?api-version=2024-09-01"
@@ -588,28 +574,6 @@ class TestAzurePromptShieldTransform:
         # Check request body
         assert call_args[1]["json"]["userPrompt"] == "test prompt"
         assert call_args[1]["json"]["documents"] == ["test prompt"]
-
-    def test_missing_http_client_raises_error(self) -> None:
-        """Missing http_client in context raises RuntimeError."""
-        from elspeth.plugins.context import PluginContext
-        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
-
-        transform = AzurePromptShield(
-            {
-                "endpoint": "https://test.cognitiveservices.azure.com",
-                "api_key": "test-key",
-                "fields": ["prompt"],
-                "schema": {"fields": "dynamic"},
-            }
-        )
-
-        ctx = Mock(spec=PluginContext, run_id="test-run")
-        ctx.http_client = None
-
-        row = {"prompt": "test", "id": 1}
-
-        with pytest.raises(RuntimeError, match="http_client"):
-            transform.process(row, ctx)
 
     def test_close_is_noop(self) -> None:
         """Close method is a no-op but exists."""
@@ -625,4 +589,420 @@ class TestAzurePromptShieldTransform:
         )
 
         # Should not raise
+        transform.close()
+
+
+class TestPromptShieldPoolConfig:
+    """Tests for Prompt Shield pool configuration."""
+
+    def test_pool_size_default_is_one(self) -> None:
+        """Default pool_size is 1 (sequential)."""
+        from elspeth.plugins.transforms.azure.prompt_shield import (
+            AzurePromptShieldConfig,
+        )
+
+        cfg = AzurePromptShieldConfig.from_dict(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+            }
+        )
+
+        assert cfg.pool_size == 1
+
+    def test_pool_size_configurable(self) -> None:
+        """pool_size can be configured."""
+        from elspeth.plugins.transforms.azure.prompt_shield import (
+            AzurePromptShieldConfig,
+        )
+
+        cfg = AzurePromptShieldConfig.from_dict(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 5,
+            }
+        )
+
+        assert cfg.pool_size == 5
+
+    def test_pool_config_property_returns_none_when_sequential(self) -> None:
+        """pool_config returns None when pool_size=1."""
+        from elspeth.plugins.transforms.azure.prompt_shield import (
+            AzurePromptShieldConfig,
+        )
+
+        cfg = AzurePromptShieldConfig.from_dict(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 1,
+            }
+        )
+
+        assert cfg.pool_config is None
+
+    def test_pool_config_property_returns_config_when_pooled(self) -> None:
+        """pool_config returns PoolConfig when pool_size>1."""
+        from elspeth.plugins.transforms.azure.prompt_shield import (
+            AzurePromptShieldConfig,
+        )
+
+        cfg = AzurePromptShieldConfig.from_dict(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 3,
+            }
+        )
+
+        assert cfg.pool_config is not None
+        assert cfg.pool_config.pool_size == 3
+
+
+class TestPromptShieldPooledExecution:
+    """Tests for Prompt Shield pooled execution."""
+
+    def test_batch_aware_is_true_when_pooled(self) -> None:
+        """Transform is batch_aware when pool_size > 1."""
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 3,
+            }
+        )
+        assert transform.is_batch_aware is True
+
+    def test_batch_aware_is_false_when_sequential(self) -> None:
+        """Transform is not batch_aware when pool_size=1."""
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 1,
+            }
+        )
+        assert transform.is_batch_aware is False
+
+    def test_pooled_execution_processes_batch_concurrently(self) -> None:
+        """Pooled transform processes batch rows concurrently."""
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 3,
+            }
+        )
+
+        # Create mock HTTP client that returns success for all calls
+        response_mock = MagicMock()
+        response_mock.status_code = 200
+        response_mock.json.return_value = {
+            "userPromptAnalysis": {"attackDetected": False},
+            "documentsAnalysis": [{"attackDetected": False}],
+        }
+        response_mock.raise_for_status = MagicMock()
+
+        mock_client = MagicMock()
+        mock_client.post.return_value = response_mock
+        transform._http_client = mock_client
+
+        # Create mock context with required landscape and state_id
+        mock_recorder = MagicMock()
+        ctx = make_mock_context()
+        ctx.landscape = mock_recorder
+        ctx.state_id = "test-state-123"
+
+        # Invoke on_start to capture recorder
+        transform.on_start(ctx)
+
+        # Process batch of 3 rows
+        rows = [
+            {"prompt": "row 1", "id": 1},
+            {"prompt": "row 2", "id": 2},
+            {"prompt": "row 3", "id": 3},
+        ]
+        result = transform.process(rows, ctx)
+
+        # Should return success_multi with all rows
+        assert result.status == "success"
+        assert result.rows is not None
+        assert len(result.rows) == 3
+
+        # Verify all rows processed
+        for i, row in enumerate(result.rows):
+            assert row["id"] == i + 1
+            assert f"row {i + 1}" in row["prompt"]
+
+        # Cleanup
+        transform.close()
+
+    def test_pooled_execution_handles_mixed_results(self) -> None:
+        """Pooled execution correctly tracks errors per row."""
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 3,
+            }
+        )
+
+        # Return attack detected based on request content
+        # (matching "malicious" keyword in the prompt)
+        def mock_post(*args: Any, **kwargs: Any) -> MagicMock:
+            response = MagicMock()
+            response.status_code = 200
+            response.raise_for_status = MagicMock()
+
+            # Extract prompt from request to determine response
+            request_json = kwargs.get("json", {})
+            prompt = request_json.get("userPrompt", "")
+
+            # Return attack detected if prompt contains "malicious"
+            if "malicious" in prompt:
+                response.json.return_value = {
+                    "userPromptAnalysis": {"attackDetected": True},
+                    "documentsAnalysis": [{"attackDetected": False}],
+                }
+            else:
+                response.json.return_value = {
+                    "userPromptAnalysis": {"attackDetected": False},
+                    "documentsAnalysis": [{"attackDetected": False}],
+                }
+            return response
+
+        mock_client = MagicMock()
+        mock_client.post.side_effect = mock_post
+        transform._http_client = mock_client
+
+        # Create mock context
+        mock_recorder = MagicMock()
+        ctx = make_mock_context()
+        ctx.landscape = mock_recorder
+        ctx.state_id = "test-state-123"
+
+        # Invoke on_start
+        transform.on_start(ctx)
+
+        # Process batch
+        rows = [
+            {"prompt": "safe 1", "id": 1},
+            {"prompt": "malicious", "id": 2},
+            {"prompt": "safe 2", "id": 3},
+        ]
+        result = transform.process(rows, ctx)
+
+        # Should return success since not all rows failed
+        assert result.status == "success"
+        assert result.rows is not None
+        assert len(result.rows) == 3
+
+        # Check that error is embedded per-row based on content
+        # Results maintain input order - find the malicious row
+        safe_count = 0
+        error_count = 0
+        for row in result.rows:
+            if "_prompt_shield_error" in row:
+                error_count += 1
+                # Verify the error row is the one with "malicious" prompt
+                assert row["prompt"] == "malicious"
+                assert row["_prompt_shield_error"]["reason"] == "prompt_injection_detected"
+            else:
+                safe_count += 1
+
+        assert safe_count == 2
+        assert error_count == 1
+
+        # Cleanup
+        transform.close()
+
+    def test_on_start_captures_recorder(self) -> None:
+        """on_start captures recorder reference for pooled execution."""
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 3,
+            }
+        )
+
+        mock_recorder = MagicMock()
+        ctx = make_mock_context()
+        ctx.landscape = mock_recorder
+
+        # Before on_start, recorder should be None
+        assert transform._recorder is None
+
+        # After on_start, recorder should be captured
+        transform.on_start(ctx)
+        assert transform._recorder is mock_recorder
+
+    def test_close_shuts_down_executor(self) -> None:
+        """close() properly shuts down the pooled executor."""
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 3,
+            }
+        )
+
+        # Executor should exist for pooled mode
+        assert transform._executor is not None
+
+        # Close should shutdown executor
+        transform.close()
+
+        # After close, recorder should be cleared
+        assert transform._recorder is None
+
+    def test_sequential_mode_has_no_executor(self) -> None:
+        """Sequential mode (pool_size=1) has no executor."""
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 1,
+            }
+        )
+
+        assert transform._executor is None
+
+    def test_audit_trail_records_api_calls(self) -> None:
+        """API calls are recorded to audit trail."""
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 3,
+            }
+        )
+
+        # Create mock HTTP client
+        response_mock = MagicMock()
+        response_mock.status_code = 200
+        response_mock.json.return_value = {
+            "userPromptAnalysis": {"attackDetected": False},
+            "documentsAnalysis": [{"attackDetected": False}],
+        }
+        response_mock.raise_for_status = MagicMock()
+
+        mock_client = MagicMock()
+        mock_client.post.return_value = response_mock
+        transform._http_client = mock_client
+
+        # Create mock recorder to track record_call invocations
+        mock_recorder = MagicMock()
+        ctx = make_mock_context()
+        ctx.landscape = mock_recorder
+        ctx.state_id = "test-state-123"
+
+        # Invoke on_start
+        transform.on_start(ctx)
+
+        # Process batch
+        rows = [
+            {"prompt": "test 1", "id": 1},
+            {"prompt": "test 2", "id": 2},
+        ]
+        transform.process(rows, ctx)
+
+        # Verify record_call was invoked for each API call
+        # Each row should have generated one call
+        assert mock_recorder.record_call.call_count == 2
+
+        # Cleanup
+        transform.close()
+
+    def test_all_rows_failed_returns_error(self) -> None:
+        """When all rows fail, returns TransformResult.error."""
+
+        from elspeth.plugins.transforms.azure.prompt_shield import AzurePromptShield
+
+        transform = AzurePromptShield(
+            {
+                "endpoint": "https://test.cognitiveservices.azure.com",
+                "api_key": "test-key",
+                "fields": ["prompt"],
+                "schema": {"fields": "dynamic"},
+                "pool_size": 3,
+            }
+        )
+
+        # Create mock HTTP client that always returns attack detected
+        response_mock = MagicMock()
+        response_mock.status_code = 200
+        response_mock.json.return_value = {
+            "userPromptAnalysis": {"attackDetected": True},
+            "documentsAnalysis": [{"attackDetected": False}],
+        }
+        response_mock.raise_for_status = MagicMock()
+
+        mock_client = MagicMock()
+        mock_client.post.return_value = response_mock
+        transform._http_client = mock_client
+
+        # Create mock context
+        mock_recorder = MagicMock()
+        ctx = make_mock_context()
+        ctx.landscape = mock_recorder
+        ctx.state_id = "test-state-123"
+
+        # Invoke on_start
+        transform.on_start(ctx)
+
+        # Process batch where all rows will fail
+        rows = [
+            {"prompt": "bad 1", "id": 1},
+            {"prompt": "bad 2", "id": 2},
+        ]
+        result = transform.process(rows, ctx)
+
+        # Should return error since all rows failed
+        assert result.status == "error"
+        assert result.reason is not None
+        assert result.reason["reason"] == "all_rows_failed"
+
+        # Cleanup
         transform.close()
